Renamed captionist route.
