+++ backend/index.js
@@ -16,7 +16,7 @@ const AdminRoutes = require("./routes/admins");
 const StudentRoutes = require("./routes/students");
 const TranscriptRoutes = require("./routes/transcripts");
 const ClassRoutes = require("./routes/classes");
-const CaptionistRoutes = require("./routes/captionist");
+const CaptionistRoutes = require("./routes/captionists");
 
 // Connect to our mongoDB instance
 mongoose.connect(
@@ -52,7 +52,7 @@ app.use("/api/admins", AdminRoutes(io));
 app.use("/api/students", StudentRoutes(io));
 app.use("/api/transcripts", TranscriptRoutes(io));
 app.use("/api/classes", ClassRoutes(io));
-app.use("/api/captionist", CaptionistRoutes(io));
+app.use("/api/captionists", CaptionistRoutes(io));
 
 // Index Route
 app.get("*", (req, res) => {
